+++ src/c_cmds.c
@@ -2422,6 +2422,7 @@ void kill_cmd_func2(char *cmd, char *parms)
     else
     {
         int kills = 0;
+        int prevkills = totalkills;
 
         if (M_StringCompare(parm, "all") || M_StringCompare(parm, "monsters"))
         {
@@ -2476,8 +2477,8 @@ void kill_cmd_func2(char *cmd, char *parms)
 
             if (kills)
             {
-                M_snprintf(buffer, sizeof(buffer), "%s%s monster%s in this map %s been killed.", (kills == 1 ? "" : "All "),
-                    commify(kills), (kills == 1 ? "" : "s"), (kills == 1 ? "has" : "have"));
+                M_snprintf(buffer, sizeof(buffer), "%s%s %smonster%s in this map %s been killed.", (kills == 1 ? "" : "All "),
+                    commify(kills), (kills < prevkills ? "remaining " : ""), (kills == 1 ? "" : "s"), (kills == 1 ? "has" : "have"));
                 C_Output(buffer);
                 C_HideConsole();
                 HU_SetPlayerMessage(buffer, false, false);