+++ releasenotes.md
@@ -1,4 +1,4 @@
-### DOOM Retro v2.7.4
+### DOOM Retro v2.8
 
 * Minor changes have been made to text that is output to the console.
 * The fuzz effect of spectres, and the player’s weapon when they have a partial invisibility power-up, are now paused when freeze mode is on.